Example file crash.
The example file calls Cli::autocomplete with the full list of words.
While this works when the input is empty, the code crash if some
characters were added.

This fix that issue by filtering the word list before calling
Cli::autocomplete.
This behavior was also added to the documentation for autocomplete
function.