Ctrl+D support
When hitting ctrl+d, tokiocli get action receive 0x04 (EOT). This caused
an issue because that character was then added to the command, messing
with autocompletion and command execution.
On top of that, the usual behavior on a ctrl+d is usually to close the program
if the line is empty.

To allow for applications to decide what to do with the ctrl+d, I've
added a NoAction value to getaction return, only when EOT is read while
the cmd is empty.
If EOT is read while cmd is non-empty, it's ignored so not to mess with
autocompletion.